Sec 116: Protecting Resources with Authorization

+++ backend/routes/posts.js
@@ -65,7 +65,7 @@ router.get("", (req, res, next) => {
   postQuery
     .then(documents => {
       fetchedPosts = documents;
-      return Post.count();
+      return Post.countDocuments();
     })
     .then(count => {
       res.status(200).json({
@@ -77,8 +77,11 @@ router.get("", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    res.status(200).json({ message: "Post deleted" });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({ message: "Deletion successful!" });
+    }
+    res.status(401).json({ message: "Not authorized!"});
   });
 });
 
@@ -99,9 +102,11 @@ router.put(
       content: req.body.content,
       imagePath: imagePath
     });
-    console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then(result => {
-      res.status(200).json({ message: "Update successful!" });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+      if (result.nModified > 0) {
+        res.status(200).json({ message: "Update successful!" });
+      }
+      res.status(401).json({ message: "Not authorized!"});
     });
   }
 );
